Remove duplicate code and refactor jobtree to toil.
