--- jobtree/src/cgcloud/jobtree/__init__.py
@@ -1,5 +1,0 @@
-from cgcloud.jobtree.jobtree_box import JobTreeBox, JobTreeLeader, JobTreeWorker
-from cgcloud.jobtree.jobtree_cluster import CreateJobTreeCluster
-
-BOXES = [ JobTreeBox, JobTreeLeader, JobTreeWorker ]
-COMMANDS = [ CreateJobTreeCluster ]
