+++ toil/setup.py
@@ -3,13 +3,13 @@ from setuptools import setup, find_packages
 cgcloud_version = '1.0.dev9'
 
 setup(
-    name='cgcloud-jobtree',
+    name='cgcloud-toil',
     version=cgcloud_version,
 
     author='Christopher Ketchum',
     author_email='cketchum@ucsc.edu',
     url='https://github.com/BD2KGenomics/cgcloud',
-    description='Setup and manage a jobtree and Apache Mesos cluster in EC2',
+    description='Setup and manage a toil and Apache Mesos cluster in EC2',
 
     package_dir={ '': 'src' },
     packages=find_packages( 'src' ),
